% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{add_administration}
\alias{add_administration}
\title{Add administration to an existing schema}
\usage{
add_administration(
  advanced_protocol,
  schema_name,
  administration,
  formulation_key = NULL
)
}
\arguments{
\item{advanced_protocol}{The protocol to which to add to the simple protocol to the schema \code{schema_name}}

\item{schema_name}{Name of the schema to add the protocol to}

\item{administration}{The protocol to add to the schema called \code{schema_name}
Must be a \code{Protocol} object, with a \code{single} dosing interval}

\item{formulation_key}{Formulation key for mapping to formulation for oral protocol. If NULL
it will be automatically assigned in the form of "Formulation X"}
}
\value{
The updated \code{AdvancedProtocol} object.
}
\description{
Add an administration to an existing schema of an advanced protocol
}
\examples{
# Create an advanced protocol
advanced_protocol <- create_advanced_protocol("Complex regimen")

# Add a schema
advanced_protocol <- add_schema(
  advanced_protocol,
  start_time = 0,
  start_time_unit = "h",
  rep_nb = 3,
  time_btw_rep = 24,
  time_btw_rep_unit = "h",
  schema_name = "Daily dosing"
)

# Create a single dose protocol
single_dose <- create_protocol(
  name = "Morning dose",
  type = "oral",
  interval = "single",
  dose = 250,
  dose_unit = "mg"
)

# Add the administration to the schema
advanced_protocol <- add_administration(
  advanced_protocol,
  schema_name = "Daily dosing",
  administration = single_dose
)
}
