% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{add_compound}
\alias{add_compound}
\title{Add compound to a \code{Simulation} object}
\usage{
add_compound(simulation, compound, protocol = NULL, formulation = list())
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}) to which the compound should be added.}

\item{compound}{Name of the compound to add to the simulation}

\item{protocol}{Name of protocol used for \code{compound}.}

\item{formulation}{Formulation key/name mapping for the chosen protocol.}
}
\value{
The \code{Simulation} object with added compound.
}
\description{
This function adds a compounds to a simulation object.
}
