% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulation.R
\name{add_formulation}
\alias{add_formulation}
\title{Add a formulation to a snapshot}
\usage{
add_formulation(snapshot, formulation)
}
\arguments{
\item{snapshot}{a snapshot object.}

\item{formulation}{a formulation object created with \code{create_formulation()}.}
}
\value{
the updated snapshot with new formulation
}
\description{
Add a formulation to a snapshot
}
\examples{
\dontrun{
# Create a formulation
tablet_form <- create_formulation(
  name = "Extended release tablet",
  type = "weibull",
  dissolution_time = 360,
  dissolution_shape = 0.75
)

# Add formulation to snapshot
snapshot <- add_formulation(snapshot, tablet_form)
}
}
