% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{add_interactions}
\alias{add_interactions}
\title{Add interactions for an already used compound from a \code{Simulation} object}
\usage{
add_interactions(simulation, compound, interactions)
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}).}

\item{compound}{name of the compound for which to add interaction}

\item{interactions}{name of interactions to use}
}
\value{
The updated \code{Simulation} object
}
\description{
This function add the interactions of the specified compounds.
}
\examples{
\dontrun{
# Add a single interaction
sim <- add_interactions(sim, "Ketoconazole", "CYP3A4-inhibition")

# Add multiple interactions
sim <- add_interactions(
  sim,
  "Rifampicin",
  c("CYP3A4-induction", "P-gp-induction")
)
}
}
