% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{add_observed_data}
\alias{add_observed_data}
\title{Add an observed dataset to a \code{Snapshot} or DDI object}
\usage{
add_observed_data(snapshot, observed_data, importer_configuration = NULL, ...)
}
\arguments{
\item{snapshot}{The \code{Snapshot} or \code{DDI} object to which the observed dataset will be added.}

\item{observed_data}{Either a \code{DataSet} object or the path to an excel file containing observed data
to be added to the \code{Snapshot} or \code{DDI} object.}

\item{importer_configuration}{Either a DataImporterConfiguration object as defined in \code{{ospsuite}} or
the path to a compatible DataImporterConfiguration.}
}
\value{
The \code{Snapshot} or \code{DDI} object with the simulation added.
}
\description{
This function adds an observed dataset to a \code{Snapshot} or \code{DDI} object.
}
