% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{add_output_interval}
\alias{add_output_interval}
\title{Adds an interval to the output schema of the simulation}
\usage{
add_output_interval(simulation, start_time, end_time, resolution, unit)
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}).}

\item{start_time}{Start time of the interval in \code{unit}}

\item{end_time}{End time of the interval in \code{unit}}

\item{resolution}{resolution in points per \code{unit}}

\item{unit}{time unit for the interval.}
}
\value{
The updated \code{Simulation} object.
}
\description{
Adds an interval to the output schema of the simulation
}
