% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{add_outputs}
\alias{add_outputs}
\title{Add outputs to a already defined outputs of a \code{Simulation} object}
\usage{
add_outputs(simulation, paths)
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}).}

\item{paths}{path of the output to add for plotting}
}
\value{
The updated \code{Simulation} object.
}
\description{
Add some outputs to be used for plotting to the simulation.
This add the output to any previously set outputs.
}
\examples{
# Add a single output path
sim <- add_outputs(sim, "Organism|Venous Blood|Plasma|Ketoconazole|Concentration")

# Add multiple output paths
sim <- add_outputs(
  sim, 
  c(
    "Organism|Liver|Intracellular|Ketoconazole|Concentration",
    "Organism|Kidney|Intracellular|Ketoconazole|Concentration"
  )
)
}
