% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{add_processes}
\alias{add_processes}
\title{Add processes for an already used compound from a \code{Simulation} object}
\usage{
add_processes(simulation, compound, processes)
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}).}

\item{compound}{name of the compound for which to add interaction}

\item{processes}{name of interactions to use}
}
\value{
The updated \code{Simulation} object
}
\description{
This function add the interactions of the specified compounds.
}
\examples{
# Add a single process
sim <- add_processes(sim, "Midazolam", "Hepatic metabolism")

# Add multiple processes
sim <- add_processes(
  sim, 
  "Ketoconazole", 
  c("Hepatic metabolism", "Renal clearance")
)
}
