% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{add_protocol}
\alias{add_protocol}
\title{Add a protocol to a snapshot}
\usage{
add_protocol(snapshot, protocol)
}
\arguments{
\item{snapshot}{a snapshot object.}

\item{protocol}{a protocol object created with \code{create_protocol()}.}
}
\value{
the updated snapshot with new protocol
}
\description{
Add a protocol to a snapshot
}
