% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{add_protocol_to_schema}
\alias{add_protocol_to_schema}
\title{Add a protocol of administration to an existing schema}
\usage{
add_protocol_to_schema(
  advanced_protocol,
  protocol,
  schema_name,
  formulation_key = NULL
)
}
\arguments{
\item{advanced_protocol}{The protocol to which to add to the simple protocol to the schema \code{schema_name}}

\item{protocol}{The protocol to add to the schema called \code{schema_name}}

\item{schema_name}{Name of the schema to add the protocol to}

\item{formulation_key}{Formulation key for mapping to formulation for oral protocol. If NULL
it will be automatically assigned in the form of "Formulation X"}
}
\value{
The updated \code{AdvancedProtocol} object.
}
\description{
Add a protocol of administration to an existing schema
}
