% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{add_schema}
\alias{add_schema}
\title{Add a schema of administration to an advanced protocol}
\usage{
add_schema(
  advanced_protocol,
  start_time,
  start_time_unit,
  rep_nb,
  time_btw_rep,
  time_btw_rep_unit,
  schema_name = NULL
)
}
\arguments{
\item{advanced_protocol}{The protocol to which to add the schema}

\item{start_time}{Starting time of the schema}

\item{start_time_unit}{Time unit for \code{start_time} of the schema}

\item{rep_nb}{Number of repetitions of the schema}

\item{time_btw_rep}{Time between repetitions of the schema}

\item{time_btw_rep_unit}{Time unit for \code{time_btw_rep} of the schema}

\item{schema_name}{Name of the schema. (Optional) If not given schema name will be in the form of "Schema X"}
}
\value{
The updated \code{AdvancedProtocol} object.
}
\description{
Add a schema to an advanced protocol
}
\examples{
# Create an advanced protocol
advanced_protocol <- create_advanced_protocol("Weekly dosing")

# Add a schema for weekly administration
advanced_protocol <- add_schema(
  advanced_protocol,
  start_time = 0,
  start_time_unit = "h",
  rep_nb = 4,
  time_btw_rep = 7,
  time_btw_rep_unit = "day",
  schema_name = "Weekly dosing"
)
}
