% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{calculatePD}
\alias{calculatePD}
\title{PD function used for Pearl index and ovulation rate}
\usage{
calculatePD(Cavg, BL, Imax, Kd, tau, hill)
}
\arguments{
\item{Cavg}{Average concentration}

\item{BL}{Baseline value for the PD function (i.e. max reachable without progestin)}

\item{Imax}{Maximum inhibition}

\item{Kd}{Dissociation constant to use for the considered progestin.}

\item{tau}{Drug specific tau (transducer function) estimate as a measure of
intrinsic efficacy}

\item{hill}{Hill coefficient used for the PD function}
}
\value{
The value of the PD function
}
\description{
PD function used for Pearl index and ovulation rate
}
