% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_internet}
\alias{check_internet}
\title{Check whether the user has an internet connection}
\usage{
check_internet()
}
\value{
TRUE if the user has an internet connection
}
\description{
Check whether the user has an internet connection
}
\keyword{internal}
