% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_pk}
\alias{compare_pk}
\title{Make a comparison table across simulations of the PK parameters}
\usage{
compare_pk(
  snapshot,
  simulation_name = NULL,
  reference_simulation_name = NULL,
  molecule_name = NULL,
  pk_parameter = NULL,
  aggregation = "mean",
  digits = 3
)
}
\arguments{
\item{snapshot}{A snapshot or ddi object for which to make a comparison
table of PK result.}

\item{simulation_name}{Optional. Simulation(s) for which to print the pk
analysis result.}

\item{reference_simulation_name}{Optional. Reference simulation name against which
the other simulation(s) will be compared to.}

\item{molecule_name}{Optional. Molecule(s) for which to print the pk
analysis result.}

\item{pk_parameter}{Optional. PK parameter(s) for which to print the pk
analysis result.}

\item{aggregation}{character string either mean or median for the type of aggregation}

\item{digits}{number of significant digits to show}
}
\description{
Make a comparison table across simulations of the PK parameters
}
