% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-blocks.R
\name{compound}
\alias{compound}
\title{Get an OSP Compound Building Block}
\usage{
compound(name)
}
\arguments{
\item{name}{character string with the name of the compound.}
}
\value{
a compound building block as a list
}
\description{
Get an OSP Compound Building Block
}
\examples{
compound("Rifampicin")
}
