% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{create_advanced_protocol}
\alias{create_advanced_protocol}
\title{Create a new advanced administration protocol}
\usage{
create_advanced_protocol(name)
}
\arguments{
\item{name}{a character string representing the protocol name.}
}
\value{
an AdvancedProtocol object
}
\description{
Create a new advanced administration protocol
}
\examples{
# Create an advanced protocol for complex dosing regimen
advanced_protocol <- create_advanced_protocol("Complex dosing regimen")
}
