% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{create_ddi}
\alias{create_ddi}
\title{Create a Drug-Drug Interaction (DDI) Simulation}
\usage{
create_ddi(victim, perpetrator)
}
\arguments{
\item{victim}{The main compound in the DDI interaction, created by \code{compound()},
mandatory.}

\item{perpetrator}{Additional compounds (perpetrators) in the DDI interaction,
created by \code{compound()}. Pass one or more \code{Compound} objects.}
}
\value{
a Drug-Drug Interaction (DDI) simulation object
}
\description{
This function creates a DDI simulation by combining a main compound (victim)
with one or more additional compounds (perpetrators).
}
\examples{
\dontrun{
rifampicin <- compound("Rifampicin")
midazolam <- compound("Midazolam")
create_ddi(victim = rifampicin, perpetrator = midazolam)
}
}
