% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{create_protocol}
\alias{create_protocol}
\title{Create a new administration protocol}
\usage{
create_protocol(
  name,
  type,
  interval,
  dose,
  dose_unit = "mg",
  end_time = 24,
  end_time_unit = "h",
  ...
)
}
\arguments{
\item{name}{a character string representing the protocol name.}

\item{type}{a character string representing the protocol type. Can be one of:
\itemize{
\item oral: oral administration
\item ivb: intravenous bolus
\item iv: intravenous
}}

\item{interval}{a character string representing the administration interval. Can be one of:
\itemize{
\item single
\item 24
\item 12-12
\item 8-8-8
\item 6-6-6-6
\item 6-6-12
}}

\item{dose}{a numeric value representing the dose.}

\item{dose_unit}{a character string representing the dose unit. Default is "mg".}

\item{end_time}{a numeric value representing the end time of the protocol. Default is 24.}

\item{end_time_unit}{a character string representing the end time unit. Default is "h".}

\item{...}{additional parameters that depends on the \code{type} of administration:
\itemize{
\item for oral administrations:
\item water_vol_per_body_weight. Default is 3.5.
\item water_vol_per_body_weight_unit: Default is "ml/kg".
\item for intravenous:
\item infusion_time. Default is 0.5.
\item infusion_time_unit. Default is "h".
}}
}
\value{
a Protocol object
}
\description{
Create a new administration protocol
}
\examples{
# Create an oral administration protocol
oral_protocol <- create_protocol(
  name = "Oral dose",
  type = "oral",
  interval = "single",
  dose = 500,
  dose_unit = "mg"
)

# Create an intravenous protocol with custom infusion time
iv_protocol <- create_protocol(
  name = "IV infusion",
  type = "iv",
  interval = "12-12",
  dose = 250,
  dose_unit = "mg",
  infusion_time = 30,
  infusion_time_unit = "min"
)
}
