% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{create_simulation}
\alias{create_simulation}
\title{Create a simulation object}
\usage{
create_simulation(
  simulation_name,
  individual = list(),
  population = list(),
  victim,
  perpetrators
)
}
\arguments{
\item{simulation_name}{Name of the simulation to be created.}

\item{individual}{Name of the individual to be used in the simulation.}

\item{population}{Name of the population to be used in the simulation.}

\item{victim}{Name of the victim compound to be used in the simulation}

\item{perpetrators}{Vector of names of the compounds to be used as perpetrators in the simulation}
}
\value{
A \code{Simulation} object
}
\description{
This function creates a custom simulation.
}
\details{
Protocols used by each compounds need to be defined afterwards with set_compound_protocol().
New compound can also be added afterwards with add_compound(). Either individual or population should
be set, but not both.
}
\examples{
# Create a simulation with an individual
sim <- create_simulation(
  simulation_name = "Drug interaction study",
  individual = "Adult male",
  victim = "Midazolam",
  perpetrators = c("Ketoconazole")
)

# Create a simulation with a population
pop_sim <- create_simulation(
  simulation_name = "Population study",
  population = "European adults",
  victim = "Warfarin",
  perpetrators = c("Rifampicin")
)
}
