% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts.R
\name{cts_table}
\alias{cts_table}
\title{Convert info gathered from CTS html to table format.}
\usage{
cts_table(info)
}
\arguments{
\item{info}{The information returned by the scraping at the
beginning of `get_cts`.}
}
\description{
Convert info gathered from CTS html to table format.
}
