% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataSetToSnapshot}
\alias{dataSetToSnapshot}
\title{Convert a DataSet object to an observed data building block}
\usage{
dataSetToSnapshot(dataset)
}
\arguments{
\item{dataset}{A DataSet object}
}
\value{
A json representation of the DataSet object
}
\description{
Convert a DataSet object to an observed data building block
}
\keyword{internal}
