% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{export_ddi}
\alias{export_ddi}
\title{Export Drug-Drug Interaction (DDI) Simulation to Snapshot}
\usage{
export_ddi(ddi, path)
}
\arguments{
\item{ddi}{a DDI object}

\item{path}{a character string that is the path to the snapshot}
}
\description{
Export Drug-Drug Interaction (DDI) Simulation to Snapshot
}
\examples{
# Create a DDI simulation
midazolam <- compound("Midazolam")
ketoconazole <- compound("Ketoconazole")
ddi <- create_ddi(victim = midazolam, ketoconazole)

# Export the DDI to a file
export_ddi(ddi, "output/Midazolam-Ketoconazole-DDI.json")
}
