% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{extract_interactions}
\alias{extract_interactions}
\title{Extract defined interactions for compounds from a snapshot}
\usage{
extract_interactions(snapshot, compounds = NULL, quietly = FALSE)
}
\arguments{
\item{snapshot}{A snapshot object.}

\item{compounds}{(Optional) A character vector of compound names to extract interactions for.
By default interactions are extracted for all compounds in the snapshot}

\item{quietly}{(Optional) Logical. If TRUE, the interactions are not printed to the console.}
}
\description{
Extract defined interactions for compounds from a snapshot
}
