% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{extract_processes}
\alias{extract_processes}
\title{Extract defined processes for compounds from a snapshot}
\usage{
extract_processes(snapshot, compounds = NULL, quietly = FALSE)
}
\arguments{
\item{snapshot}{A snapshot object.}

\item{compounds}{(Optional) A character vector of compound names to extract processes for.
By default processes are extracted for all compounds in the snapshot}

\item{quietly}{(Optional) Logical. If TRUE, the processes are not printed to the console.}
}
\description{
Extract defined processes for compounds from a snapshot
}
