% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenCurves.R
\name{gen_Sample}
\alias{gen_Sample}
\title{Generate simulated sample data}
\usage{
gen_Sample(
  N = 1000,
  meanPar = list(c(log(0.2), log(5)), c(log(0.22), log(4.5))),
  sdPar = list(c(log(0.5), log(0.5)), c(log(3), log(3))),
  dose = 100,
  iov.sd = 0.5,
  seed = -17
)
}
\arguments{
\item{N}{number of profiles per condition}

\item{meanPar}{List of length two. Each item is a vector with the log10(mean)
for Ke and V.  The first item is the reference compound; the second item is the
generic.}

\item{sdPar}{List of length two. Each item is a vector with the log10(sd)
for Ke and V.  The first item is the reference compound; the second item is the
generic.}

\item{dose}{Simulated dose for both reference and generic compounds}

\item{iov.sd}{SD to apply for random intra-occasion variability in parameter values.}

\item{seed}{Seed for random number generator.}
}
\value{
A tibble with columns ID, time, conc, per, drug
}
\description{
Creates concentration time curves for clinical trial simulator examples from a
one-compartment model with instantaneous bolus input.
}
\author{
Michael Neely
}
