% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{get_AUC}
\alias{get_AUC}
\title{Calculate AUC}
\usage{
get_AUC(x)
}
\arguments{
\item{x}{An appropriate tibble.}
}
\value{
A tibble with id and auc.
}
\description{
Calculate AUC using trapezoidal approximation from tibble read
by \code{\link{get_data}}.
}
\author{
Michael Neely
}
