% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{get_Cmax}
\alias{get_Cmax}
\title{Extract Cmax}
\usage{
get_Cmax(x)
}
\arguments{
\item{x}{An appropriate tibble.}
}
\value{
A tibble with id and cmax.
}
\description{
Extract Cmax from tibble read by \code{\link{get_data}}.
}
\author{
Michael Neely
}
