% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_cts}
\alias{get_cts}
\title{Get the available "Congress Trump Score" information for each congressperson.}
\usage{
get_cts(chamber = "both")
}
\arguments{
\item{chamber}{Chamber from which to get the information. Can assume
one of 3 values: `house`, `senate`, or `both`.}
}
\description{
Get the available "Congress Trump Score" information for each congressperson.
}
