% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{get_data}
\alias{get_data}
\title{Read data file for BE simulations}
\usage{
get_data(name)
}
\arguments{
\item{name}{Name of the data file.}
}
\value{
A tibble.
}
\description{
Read a csv file whose format should be subject id, time, conc, period, drug.
}
\author{
Michael Neely
}
