% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{get_source}
\alias{get_source}
\title{Get the source file path or url from the input.}
\usage{
get_source(input)
}
\arguments{
\item{input}{character string that is wether
\itemize{
\item a compound name from the list of available compounds \code{list_compounds()}.
\item a URL to a compound building block.
\item a Path to a local file.
}}
}
\value{
The source file path or url.
}
\description{
Get the source file path or url from the input.
}
