% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts.R
\name{get_votes}
\alias{get_votes}
\title{Get the vote data for each congressperson}
\usage{
get_votes(chamber = "both", progress = TRUE)
}
\arguments{
\item{chamber}{Chamber from which to get the information. Can assume
one of 3 values: `house`, `senate`, or `both`.}

\item{progress}{Whether to show a progress bar when running.}
}
\description{
Get the vote data for each congressperson
}
