% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{hline}
\alias{hline}
\title{Add horizontal reference line to plotly}
\usage{
hline(y = 0, color = "red", dash = "solid")
}
\arguments{
\item{y}{Y-intercept}

\item{color}{Color}

\item{dash}{Character setting the dash style of lines. Set to a dash type string (solid,
dot, dash, longdash, dashdot, or longdashdot) or a dash length list in px
(eg 5px,10px,2px,2px).}
}
\value{
A line
}
\description{
Add horizontal reference line to plotly
}
\author{
Michael Neely
}
