% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_file_local}
\alias{is_file_local}
\title{Check if the input is a file}
\usage{
is_file_local(x)
}
\arguments{
\item{x}{string, the input to check}
}
\value{
TRUE if the input is a file
}
\description{
Check if the input is a file
}
\keyword{internal}
