% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_file_url}
\alias{is_file_url}
\title{Check if the input is a URL to a file}
\usage{
is_file_url(x, error_call = caller_env())
}
\arguments{
\item{x}{string, the input to check}
}
\value{
TRUE if the input is a URL
}
\description{
Check if the input is a URL to a file
}
\keyword{internal}
