% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building-blocks.R
\name{list_compounds}
\alias{list_compounds}
\title{Available Compound Building Blocks}
\usage{
list_compounds(...)
}
\value{
a list of the available compound building blocks
}
\description{
List all available official compound building blocks on OSPsuite
}
\examples{
list_compounds()
}
