% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{make_report}
\alias{make_report}
\title{Generate BE report}
\usage{
make_report(sim, be)
}
\arguments{
\item{sim}{Raw simulated data from \code{\link{get_data}}.}

\item{be}{Output of \code{\link{extractBE}}}
}
\value{
HTML document will open in browser
}
\description{
Generate BE report
}
\author{
Michael Neely
}
