% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{ovulationRate}
\alias{ovulationRate}
\title{Calculate ovulation rate for a DDI object}
\usage{
ovulationRate(ddi, simulations = NULL, Kd = NULL, tau = NULL)
}
\arguments{
\item{ddi}{ddi snapshot object}

\item{simulations}{simulations names for which to calculate pearl index}

\item{Kd}{Dissociation constant to use for the considered progestin
(only Kd for DRSP and LNG are known). If given the value superseed the default
values for LNG and DRSP.}

\item{tau}{Drug specific tau (transducer function) estimate as a measure of
intrinsic efficacy}
}
\value{
a list of ovulation rate for each required simulation.
}
\description{
This function calculate Ovulation Rate for the ddi victim (i.e. progestin) based on
the simulation average concentration at the end of the simulation defined as
$1/2 (C_{max} + C_{through})$ for the last dosing of the simulation.
This should only be used for simulations meaningful repeated administration.
}
