% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_pk_analysis}
\alias{pivot_pk_analysis}
\title{Format PK Analysis Data}
\usage{
pivot_pk_analysis(df, compound_names)
}
\arguments{
\item{df}{A data frame or tibble containing PK analysis data returned by
\code{ospsuite::calculatePKAnalyses}.}

\item{compound_names}{A character vector of compound names.}
}
\value{
A tibble with individual columns for each compound.
}
\description{
Transforms PK analysis data by removing compound names from \code{QuantityPath},
identifying compounds, and pivoting to a wide format with columns for each
compound.
}
\keyword{internal}
