% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{plot_ddi_results}
\alias{plot_ddi_results}
\title{Plot DDI simulations defined in the ddi project}
\usage{
plot_ddi_results(ddi, simulationNames = NULL, plotConfiguration = NULL)
}
\arguments{
\item{ddi}{a DDI object}

\item{simulationNames}{a character vector of simulation names for which to generate plots.
Default is NULL, i.e. plots will be generated for all simulations.}

\item{plotConfiguration}{a DefaultPlotConfiguration object to customize the plots.
If a yUnit is provided, only plots for that unit will be generated.}
}
\value{
a list of plots
}
\description{
Plot DDI simulations defined in the ddi project
}
