% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBE.R
\name{plotBE}
\alias{plotBE}
\title{Plot BE results}
\usage{
plotBE(x, type, ci_lo = 0.8, ci_up = 1.25, ylab)
}
\arguments{
\item{x}{Output of \code{\link{extractBE}} to plot.}

\item{ci_lo}{Lower confidence interval reference line for geometric mean AUC or Cmax to be
considered BE, default 0.8 per FDA guidance.}

\item{ci_up}{Upper confidence interval reference line for geometric mean AUC or Cmax to be
considered BE, default 1.25 per FDA guidance.}

\item{ylab}{Label for y-axis to override defaults.}
}
\value{
A plot.
}
\description{
Generates two types of plots: ratio and probability of BE for given sample size.
Ratio plots are the geometric mean by sample size of either AUC or Cmax and
confidence interval, as calculated by \code{\link{extractBE}}.
Probability plots are the probability of demonstrating BE by sample size.
}
\author{
Michael Neely
}
