% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBE.R
\name{plotQ}
\alias{plotQ}
\title{Plot quantiles of simulated data}
\usage{
plotQ(x, ci = 0.9)
}
\arguments{
\item{x}{Output of \code{\link{get_data}}}

\item{ci}{Confidence interval to plot, default is 0.9 (90\%)
per FDA guidance.}

\item{ylab}{Label for y-axis to override defaults.}
}
\value{
A plot.
}
\description{
Generates the median and lower/upper quantiles of the simulated data,
stratified by drug.
}
\author{
Michael Neely
}
