% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pretty_pk}
\alias{pretty_pk}
\title{Pretty print PK Analysis Data}
\usage{
pretty_pk(
  snapshot,
  simulation_name = NULL,
  molecule_name = NULL,
  pk_parameter = NULL,
  aggregation = "mean",
  digits = 3
)
}
\arguments{
\item{snapshot}{A snapshot or ddi object for which to make a pretty table of PK result}

\item{simulation_name}{Optional for which simulation(s) to print the pk analysis result in pretty format}

\item{molecule_name}{Optional for which molecule(s) to print the pk analysis result in pretty format}

\item{pk_parameter}{Optional for which PK parameters to print the pk analysis result in pretty format}

\item{aggregation}{character string either mean or median for the type of aggregation}

\item{digits}{number of significant digits to show}
}
\description{
Print PK analysis data to a pretty table
}
