% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{remove_administration}
\alias{remove_administration}
\title{Remove administration from an existing schema}
\usage{
remove_administration(advanced_protocol, schema_name, protocol_name)
}
\arguments{
\item{advanced_protocol}{The protocol to which to remove the simple protocol from the schema \code{schema_name}}

\item{schema_name}{Name of the schema to remove the protocol from}

\item{protocol}{The protocol to remove from the schema called \code{schema_name}}
}
\value{
The updated \code{AdvancedProtocol} object.
}
\description{
Remove an administration from an existing schema of an advanced protocol
}
