% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{remove_observed_data}
\alias{remove_observed_data}
\title{Remove observed dataset(s) from a \code{Snapshot} or DDI object}
\usage{
remove_observed_data(snapshot, observed_data_name)
}
\arguments{
\item{snapshot}{The \code{Snapshot} or \code{DDI} object to which the observed dataset(s) should be removed.}

\item{observed_data_name}{Name(s) of the observed datasets to be removed from the \code{Snapshot} or \code{DDI} object}
}
\value{
The \code{Snapshot} or \code{DDI} object with the observed dataset(s) removed
}
\description{
This function removes an observed dataset from a \code{Snapshot} or \code{DDI} object.
}
