% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{remove_protocol}
\alias{remove_protocol}
\title{Remove a protocol from a snapshot}
\usage{
remove_protocol(snapshot, protocol_name)
}
\arguments{
\item{snapshot}{a snapshot object.}

\item{protocol_name}{a character vector of protocol name(s) to remove}
}
\value{
the updated snapshot without the specified protocols
}
\description{
Remove a protocol from a snapshot
}
