% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{remove_schema}
\alias{remove_schema}
\title{Remove a schema of administration from an advanced protocol}
\usage{
remove_schema(advanced_protocol, schema_name)
}
\arguments{
\item{advanced_protocol}{The protocol to which to remove the schema}

\item{schema_name}{Name of the schema to remove}
}
\value{
The updated \code{AdvancedProtocol} object.
}
\description{
Remove a schema from an advanced protocol
}
\examples{
#' # Create an advanced protocol
advanced_protocol <- create_advanced_protocol("Weekly dosing")

# Add a schema for weekly administration
advanced_protocol <- add_schema(
  advanced_protocol,
  start_time = 0,
  start_time_unit = "h",
  rep_nb = 4,
  time_btw_rep = 7,
  time_btw_rep_unit = "day",
  schema_name = "Weekly dosing"
)

# Remove a schema from an advanced protocol
advanced_protocol <- remove_schema(advanced_protocol, "Weekly dosing")
}
