% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{remove_simulation}
\alias{remove_simulation}
\title{Remove simulation(s) from a \code{Snapshot} or \code{DDI} object}
\usage{
remove_simulation(snapshot, simulation_name)
}
\arguments{
\item{snapshot}{The \code{Snapshot} or \code{DDI} object to which the simulation(s) should be removed.}

\item{simulation_names}{Names of the simulations to be removed from the \code{Snapshot} or \code{DDI} object}
}
\value{
The \code{Snapshot} or \code{DDI} object with the simulation(s) removed
}
\description{
This function removes a simulation from a \code{Snapshot} or \code{DDI} object.
}
