% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{run_ddi}
\alias{run_ddi}
\title{Run Drug-Drug Interaction (DDI) Simulations defined in the ddi project}
\usage{
run_ddi(ddi, path = NULL, exportPKML = FALSE)
}
\arguments{
\item{ddi}{a DDI object}

\item{path}{description}

\item{exportPKML}{description}
}
\description{
Run Drug-Drug Interaction (DDI) Simulations defined in the ddi project
}
