% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{run_pk_analysis}
\alias{run_pk_analysis}
\title{Run Pk-Analysis for DDI simulations defined in the ddi project}
\usage{
run_pk_analysis(ddi, path = NULL, exportToCSV = FALSE)
}
\arguments{
\item{ddi}{a DDI object}

\item{path}{description}

\item{exportToCSV}{description}
}
\description{
Run Pk-Analysis for DDI simulations defined in the ddi project
}
