% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{set_compound_protocol}
\alias{set_compound_protocol}
\title{Set Protocol and Formulation for a Compound in a Simulation}
\usage{
set_compound_protocol(simulation, compound, protocol, formulation = list())
}
\arguments{
\item{simulation}{A Simulation object}

\item{compound}{Name of the compound}

\item{protocol}{A character string specifying the protocol name}

\item{formulation}{Optional. Either a character string or a character vector of formulation names.
If a single formulation is provided, it will be used for all protocol keys.
If multiple formulations are provided, they will be mapped in order to the protocol keys.}
}
\value{
The simulation object (invisibly)
}
\description{
Sets the protocol and formulation for a compound in a simulation.
}
\examples{
sim <- create_simulation(
  simulation_name = "Test",
  victim = "Rifampicin",
  perpetrators = "Midazolam",
  individual = "European (P-gp modified, CYP3A4 36 h)"
)
sim <- set_compound_protocol(sim, "Midazolam", "Single oral dose", "Tablet")

}
