% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{set_output_interval}
\alias{set_output_interval}
\title{Clears the output interval from the simulation and adds a new one.}
\usage{
set_output_interval(simulation, start_time, end_time, resolution, unit)
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}).}

\item{start_time}{Start time of the interval in \code{unit}}

\item{end_time}{End time of the interval in \code{unit}}

\item{resolution}{resolution in points per \code{unit}}

\item{unit}{time unit for the interval. Should be one of ospsuite::ospUnits$Time.}
}
\value{
The updated \code{Simulation} object.
}
\description{
Clears the output interval from the simulation and adds a new one.
}
\examples{
# Create a simulation first
sim <- create_simulation(
  simulation_name = "Output interval example",
  individual = "Adult male",
  victim = "Midazolam",
  perpetrators = c("Ketoconazole")
)

# Set output interval for 24 hours with 10 points per hour
sim <- set_output_interval(sim, 0, 24, 10, "h")

# Set output interval for 7 days with 24 points per day
sim <- set_output_interval(sim, 0, 7, 24, "day(s)")
}
