% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{set_outputs}
\alias{set_outputs}
\title{Set outputs to a \code{Simulation} object}
\usage{
set_outputs(simulation, paths)
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}).}

\item{paths}{path of the output to use for plotting}
}
\value{
The updated \code{Simulation} object.
}
\description{
Set the outputs to be used for plotting to the simulation.
This removes any previously set outputs.
}
\examples{
# Create a simulation first
sim <- create_simulation(
  simulation_name = "Output example",
  individual = "Adult male",
  victim = "Midazolam",
  perpetrators = c("Ketoconazole")
)

# Set a single output path
sim <- set_outputs(sim, "Organism|Venous Blood|Plasma|Midazolam|Concentration")

# Set multiple output paths
sim <- set_outputs(
  sim,
  c(
    "Organism|Venous Blood|Plasma|Midazolam|Concentration",
    "Organism|Liver|Intracellular|Midazolam|Concentration"
  )
)
}
