% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputSchema.R
\name{SnapshotOutputSchema}
\alias{SnapshotOutputSchema}
\title{R6 Class Representing a Snapshot Output Schema}
\description{
This class is used to create the json corresponding to a simulation output schema snapshot
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{intervals}}{All intervals defined in the schema (Read-Only)}

    \item{\code{data}}{dynamic json representation of the output schema object}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-SnapshotOutputSchema-clear}{\code{SnapshotOutputSchema$clear()}}
    \item \href{#method-SnapshotOutputSchema-add_interval}{\code{SnapshotOutputSchema$add_interval()}}
    \item \href{#method-SnapshotOutputSchema-set_interval}{\code{SnapshotOutputSchema$set_interval()}}
    \item \href{#method-SnapshotOutputSchema-print}{\code{SnapshotOutputSchema$print()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnapshotOutputSchema-clear"></a>}}
\if{latex}{\out{\hypertarget{method-SnapshotOutputSchema-clear}{}}}
\subsection{\code{SnapshotOutputSchema$clear()}}{
  Clears all intervals
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SnapshotOutputSchema$clear()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnapshotOutputSchema-add_interval"></a>}}
\if{latex}{\out{\hypertarget{method-SnapshotOutputSchema-add_interval}{}}}
\subsection{\code{SnapshotOutputSchema$add_interval()}}{
  Adds an interval to the schema
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SnapshotOutputSchema$add_interval(start_time, end_time, resolution, unit = "h")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start_time}}{Start time for the interval in \code{unit}}
      \item{\code{end_time}}{End time for the interval in \code{unit}}
      \item{\code{resolution}}{resolution in points per \code{unit}}
      \item{\code{unit}}{time unit for the interval. Should be one of ospsuite::ospUnits$Time.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnapshotOutputSchema-set_interval"></a>}}
\if{latex}{\out{\hypertarget{method-SnapshotOutputSchema-set_interval}{}}}
\subsection{\code{SnapshotOutputSchema$set_interval()}}{
  Clear any previously existing intervals and set a new one
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SnapshotOutputSchema$set_interval(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Same parameters as add_interval}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnapshotOutputSchema-print"></a>}}
\if{latex}{\out{\hypertarget{method-SnapshotOutputSchema-print}{}}}
\subsection{\code{SnapshotOutputSchema$print()}}{
  Print the object to the console
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SnapshotOutputSchema$print()}
    \if{html}{\out{</div>}}
  }
}

}
