% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_list}
\alias{to_list}
\title{Convert an object or vector to a list}
\usage{
to_list(x)
}
\arguments{
\item{x}{Any R object or vector.}
}
\value{
A list containing \code{x}, a list of the elements of \code{x} if \code{x} is a
vector, or \code{x} itself if it's already a list.
}
\description{
If the input is already a list, it returns the input unchanged. If the input
is a vector, it converts the vector to a list where each element of the vector
becomes an element of the list. For all other types, it wraps the input in a
list.
}
\keyword{internal}
