% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cts.R
\name{to_table}
\alias{to_table}
\title{Convert info gathered from html to table format.}
\usage{
to_table(info)
}
\arguments{
\item{info}{The information returned by the scraping at the
beginning of `get_cts`.}
}
\description{
Convert info gathered from html to table format.
}
