% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts.R
\name{votes_table}
\alias{votes_table}
\title{Convert info gathered from votes html to table format.}
\usage{
votes_table(info, name)
}
\arguments{
\item{info}{The information returned by the scraping at the
central loop of `get_votes`.}

\item{name}{Name of the congressperson whose information is
being gathered.}
}
\description{
Convert info gathered from votes html to table format.
}
