% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_json_suffix}
\alias{with_json_suffix}
\title{Add .json suffix to file path(s)}
\usage{
with_json_suffix(path)
}
\arguments{
\item{path}{A character vector of file paths.}
}
\value{
A character vector with the ".json" suffix added where necessary.
}
\description{
This function checks if the provided file path(s) already have the ".json"
suffix. If not, it appends the ".json" suffix. The function is vectorized.
}
\keyword{internal}
