% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{add_administration}
\alias{add_administration}
\title{Add administration to an existing schema}
\usage{
add_administration(
  advanced_protocol,
  schema_name,
  administration,
  formulation_key = NULL
)
}
\arguments{
\item{advanced_protocol}{The protocol to which to add to the simple protocol to the schema \code{schema_name}}

\item{schema_name}{Name of the schema to add the protocol to}

\item{administration}{The protocol to add to the schema called \code{schema_name}
Must be a \code{Protocol} object, with a \code{single} dosing interval}

\item{formulation_key}{Formulation key for mapping to formulation for oral protocol. If NULL
it will be automatically assigned in the form of "Formulation X"}
}
\value{
The updated \code{AdvancedProtocol} object.
}
\description{
Add an administration to an existing schema of an advanced protocol
}
