% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{add_outputs}
\alias{add_outputs}
\title{Add outputs to a already defined outputs of a \code{Simulation} object}
\usage{
add_outputs(simulation, paths)
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}).}

\item{paths}{path of the output to add for plotting}
}
\value{
The updated \code{Simulation} object.
}
\description{
Add some outputs to be used for plotting to the simulation.
This add the output to any previously set outputs.
}
