% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{add_simulation}
\alias{add_simulation}
\title{Add a simulation to a \code{Snapshot} or DDI object}
\usage{
add_simulation(snapshot, simulation)
}
\arguments{
\item{snapshot}{The \code{Snapshot} or \code{DDI} object to which the simulation will be added.}

\item{simulation}{\code{Simulation} object created by \code{create_simulation()}, to be added to the \code{Snapshot} or \code{DDI} object.}
}
\value{
The \code{Snapshot} or \code{DDI} object with the simulation added.
}
\description{
This function adds a simulation to a \code{Snapshot} or \code{DDI} object.
}
