% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{Compound}
\alias{Compound}
\title{R6 Class Representing a Compound Snapshot}
\description{
A snapshot is a JSON file that contains all the information about a model.
This class is used to import, access and manipulate the data from a compound snapshot.
}
\section{Super class}{
\code{\link[cts:Snapshot]{cts::Snapshot}} -> \code{Compound}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Compound-clone}{\code{Compound$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="compoundsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-compoundsNames'><code>cts::Snapshot$compoundsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="eventsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-eventsNames'><code>cts::Snapshot$eventsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="export"><a href='../../cts/html/Snapshot.html#method-Snapshot-export'><code>cts::Snapshot$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="formulationsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-formulationsNames'><code>cts::Snapshot$formulationsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="individualsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-individualsNames'><code>cts::Snapshot$individualsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="initialize"><a href='../../cts/html/Snapshot.html#method-Snapshot-initialize'><code>cts::Snapshot$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="observedDataNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-observedDataNames'><code>cts::Snapshot$observedDataNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="observerSetsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-observerSetsNames'><code>cts::Snapshot$observerSetsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="populationsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-populationsNames'><code>cts::Snapshot$populationsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="print"><a href='../../cts/html/Snapshot.html#method-Snapshot-print'><code>cts::Snapshot$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="protocolsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-protocolsNames'><code>cts::Snapshot$protocolsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="simulationsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-simulationsNames'><code>cts::Snapshot$simulationsNames()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compound-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Compound-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compound$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
