% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{compound}
\alias{compound}
\title{Create a compound building block}
\usage{
compound(input)
}
\arguments{
\item{input}{character string that is wether
\itemize{
\item a compound name from the list of available compounds \code{list_compounds()}.
\item a URL to a compound building block.
\item a Path to a local file.
}}
}
\value{
a compound building block as a list
}
\description{
Create a compound building block
}
\examples{
compound("Rifampicin")
compound("https://raw.githubusercontent.com/Open-Systems-Pharmacology/Alfentanil-Model/v2.2/Alfentanil-Model.json")
\dontrun{
compound("Rifampicin-Model.json")
#'
}
}
