% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{Compound}
\alias{Compound}
\title{R6 Class Representing a Compound Snapshot}
\description{
A snapshot is a JSON file that contains all the information about a model.
This class is used to import, access and manipulate the data from a compound snapshot.
}
\section{Super class}{
\code{\link[cts:Snapshot]{cts::Snapshot}} -> \code{Compound}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Compound-clone}{\code{Compound$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="add_formulation"><a href='../../cts/html/Snapshot.html#method-Snapshot-add_formulation'><code>cts::Snapshot$add_formulation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="add_protocol"><a href='../../cts/html/Snapshot.html#method-Snapshot-add_protocol'><code>cts::Snapshot$add_protocol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="add_simulation"><a href='../../cts/html/Snapshot.html#method-Snapshot-add_simulation'><code>cts::Snapshot$add_simulation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="check_simulation"><a href='../../cts/html/Snapshot.html#method-Snapshot-check_simulation'><code>cts::Snapshot$check_simulation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="create_plots"><a href='../../cts/html/Snapshot.html#method-Snapshot-create_plots'><code>cts::Snapshot$create_plots()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="export"><a href='../../cts/html/Snapshot.html#method-Snapshot-export'><code>cts::Snapshot$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="get_names"><a href='../../cts/html/Snapshot.html#method-Snapshot-get_names'><code>cts::Snapshot$get_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="initialize"><a href='../../cts/html/Snapshot.html#method-Snapshot-initialize'><code>cts::Snapshot$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="print"><a href='../../cts/html/Snapshot.html#method-Snapshot-print'><code>cts::Snapshot$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="remove_formulation"><a href='../../cts/html/Snapshot.html#method-Snapshot-remove_formulation'><code>cts::Snapshot$remove_formulation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="remove_protocol"><a href='../../cts/html/Snapshot.html#method-Snapshot-remove_protocol'><code>cts::Snapshot$remove_protocol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="remove_simulation"><a href='../../cts/html/Snapshot.html#method-Snapshot-remove_simulation'><code>cts::Snapshot$remove_simulation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="run_pk_analysis"><a href='../../cts/html/Snapshot.html#method-Snapshot-run_pk_analysis'><code>cts::Snapshot$run_pk_analysis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="run_simulations"><a href='../../cts/html/Snapshot.html#method-Snapshot-run_simulations'><code>cts::Snapshot$run_simulations()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compound-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Compound-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compound$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
