% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{create_ddi}
\alias{create_ddi}
\title{Create a Drug-Drug Interaction (DDI) Simulation}
\usage{
create_ddi(victim, perpetrator, options = NULL)
}
\arguments{
\item{victim}{The main compound in the DDI interaction, created by \code{compound()},
mandatory.}

\item{perpetrator}{Additional compounds (perpetrators) in the DDI interaction,
created by \code{compound()}. Pass one or more \code{Compound} objects.}

\item{options}{a named list of options to customize the DDI simulation. Default is to use \code{default_options}.
\itemize{
\item import_simulations: logical, whether to import the simulations from the victim and perpetrators. default is FALSE.
\item auto_create_ddi_simulation: logical, whether to create a generic simulation template. default is TRUE.
generic simulation will use:
\itemize{
\item the first compound defined as perpetrator,
\item the first protocol of the victim and perpetrator compounds,
\item the first formulation of the victim and perpetrator compounds,
\item the first individual defined in victim compounds.
}
}}
}
\value{
a Drug-Drug Interaction (DDI) simulation object
}
\description{
This function creates a DDI simulation by combining a main compound (victim)
with one or more additional compounds (perpetrators).
}
\examples{
\dontrun{
rifampicin <- compound("Rifampicin")
midazolam <- compound("Midazolam")
create_ddi(victim = rifampicin, perpetrator = midazolam)
}
}
