% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{create_ddi}
\alias{create_ddi}
\title{Create a Drug-Drug Interaction (DDI) Simulation}
\usage{
create_ddi(...)
}
\arguments{
\item{...}{a set of compounds created by \code{compound()}}
}
\value{
a Drug-Drug Interaction (DDI) simulation object
}
\description{
Create a Drug-Drug Interaction (DDI) Simulation
}
\examples{
\dontrun{
rifampicin <- compound("Rifampicin")
midazolam <- compound("Midazolam")
create_ddi(rifampicin, midazolam)
}
}
