% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{create_ddi}
\alias{create_ddi}
\title{Create a Drug-Drug Interaction (DDI) Simulation}
\usage{
create_ddi(victim, ..., options = NULL)
}
\arguments{
\item{victim}{The main compound in the DDI interaction, created by \code{compound()},
mandatory.}

\item{...}{Additional \code{Compounds} objects to be added as perpetrators in the
DDI interaction,}

\item{options}{a named list of options to customize the DDI simulation. Default is to use \code{default_options}.
\itemize{
\item import_simulations: logical, whether to import the simulations from the victim and perpetrators. default is FALSE.
\item create_ddi_simulation: logical, whether to create a generic simulation template. default is TRUE.
generic simulation will use:
\itemize{
\item the first compound defined as perpetrator,
\item the first protocol of the victim and perpetrator compounds,
\item the first formulation of the victim and perpetrator compounds,
\item the first individual defined in victim compounds.
}
}}
}
\value{
a Drug-Drug Interaction (DDI) simulation object
}
\description{
This function creates a DDI simulation by combining a main compound (victim)
with one or more additional compounds (perpetrators).
}
\examples{
# Create compounds
midazolam <- compound("Midazolam")
ketoconazole <- compound("Ketoconazole")

# Create a basic DDI simulation
ddi <- create_ddi(victim = midazolam, ketoconazole)

# Create a DDI with custom options
ddi_custom <- create_ddi(
  victim = midazolam, 
  ketoconazole,
  options = list(
    import_simulations = TRUE,
    create_ddi_simulation = FALSE
  )
)

# Create a DDI with multiple perpetrators
rifampicin <- compound("Rifampicin")
ddi_multi <- create_ddi(victim = midazolam, ketoconazole, rifampicin)
}
