% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts-package.R
\docType{package}
\name{cts-package}
\alias{cts}
\alias{cts-package}
\title{cts: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://esqlabs.github.io/cts/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
