% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS-package.R
\docType{package}
\name{CTS-package}
\alias{CTS-package}
\alias{CTS}
\title{Clinical trial simlation for BE studies}
\description{
This package contains functions to generate statistics for various sample
sizes and trial designs based on models created by PK-SIM + Pmetrics.
}
\author{
Michael Neely MD
\url{http://www.lapk.org}
}
\keyword{package}
