% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{DDI}
\alias{DDI}
\title{R6 Class Representing a DDI Snapshot}
\description{
A snapshot is a JSON file that contains all the information about a model.
This class is used to import, access and manipulate the data from a DDI snapshot.
}
\section{Super class}{
\code{\link[cts:Snapshot]{cts::Snapshot}} -> \code{DDI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DDI-new}{\code{DDI$new()}}
\item \href{#method-DDI-combine}{\code{DDI$combine()}}
\item \href{#method-DDI-import}{\code{DDI$import()}}
\item \href{#method-DDI-clone}{\code{DDI$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="compoundsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-compoundsNames'><code>cts::Snapshot$compoundsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="eventsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-eventsNames'><code>cts::Snapshot$eventsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="export"><a href='../../cts/html/Snapshot.html#method-Snapshot-export'><code>cts::Snapshot$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="formulationsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-formulationsNames'><code>cts::Snapshot$formulationsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="individualsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-individualsNames'><code>cts::Snapshot$individualsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="observedDataNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-observedDataNames'><code>cts::Snapshot$observedDataNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="observerSetsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-observerSetsNames'><code>cts::Snapshot$observerSetsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="populationsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-populationsNames'><code>cts::Snapshot$populationsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="print"><a href='../../cts/html/Snapshot.html#method-Snapshot-print'><code>cts::Snapshot$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="protocolsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-protocolsNames'><code>cts::Snapshot$protocolsNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cts" data-topic="Snapshot" data-id="simulationsNames"><a href='../../cts/html/Snapshot.html#method-Snapshot-simulationsNames'><code>cts::Snapshot$simulationsNames()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DDI-new"></a>}}
\if{latex}{\out{\hypertarget{method-DDI-new}{}}}
\subsection{Method \code{new()}}{
Create a DDI object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDI$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{DDI} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DDI-combine"></a>}}
\if{latex}{\out{\hypertarget{method-DDI-combine}{}}}
\subsection{Method \code{combine()}}{
Combine multiple compounds into a DDI simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDI$combine(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{a set of compounds created by \code{compound()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DDI-import"></a>}}
\if{latex}{\out{\hypertarget{method-DDI-import}{}}}
\subsection{Method \code{import()}}{
Import a DDI simulation from a JSON file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDI$import(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{character string that is wether
\itemize{
\item a compound name from the list of available compounds \code{list_compounds()}.
\item a URL to a compound building block.
\item a Path to a local file.
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DDI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DDI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DDI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
