% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{import_ddi}
\alias{import_ddi}
\title{Import Drug-Drug Interaction (DDI) Simulation from Snapshot}
\usage{
import_ddi(input)
}
\arguments{
\item{input}{character string that is wether
\itemize{
\item a compound name from the list of available compounds \code{list_compounds()}.
\item a URL to a compound building block.
\item a Path to a local file.
}}
}
\value{
a DDI object
}
\description{
Import Drug-Drug Interaction (DDI) Simulation from Snapshot
}
\examples{
#' \dontrun{
import_ddi("Rifampicin-Midazolam-DDI.json")
}
}
