% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{plot_ddi_results}
\alias{plot_ddi_results}
\title{Plot DDI simulations defined in the ddi project}
\usage{
plot_ddi_results(ddi, simulationNames = NULL, ...)
}
\arguments{
\item{ddi}{a DDI object}

\item{simulationNames}{a character vector of simulation names for which to generate plots.
Default is NULL, i.e. plots will be generated for all simulations.}

\item{...}{additional arguments to pass to the plotPopulationTimeProfile (for example aggregation method)}
}
\value{
a list of plots
}
\description{
Plot DDI simulations defined in the ddi project
}
\examples{
# Create compounds
midazolam <- compound("Midazolam")
itraconazole <- compound("Itraconazole")

# Create a DDI simulation
ddi <- create_ddi(victim = midazolam, itraconazole)

# Plot all simulation results
\dontrun{
plots <- plot_ddi_results(ddi)
}
}
