% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{remove_administration}
\alias{remove_administration}
\title{Remove administration from an existing schema}
\usage{
remove_administration(advanced_protocol, schema_name, administration_name)
}
\arguments{
\item{advanced_protocol}{The protocol to which to remove the simple protocol from the schema \code{schema_name}}

\item{schema_name}{Name of the schema to remove the protocol from}

\item{administration_name}{The name of the administration protocol to remove from the schema called \code{schema_name}}
}
\value{
The updated \code{AdvancedProtocol} object.
}
\description{
Remove an administration from an existing schema of an advanced protocol
}
\examples{
# Create an advanced protocol
advanced_protocol <- create_advanced_protocol("Complex regimen")

# Add a schema
advanced_protocol <- add_schema(
  advanced_protocol,
  start_time = 0,
  start_time_unit = "h",
  rep_nb = 3,
  time_btw_rep = 24,
  time_btw_rep_unit = "h",
  schema_name = "Daily dosing"
)

# Create and add a single dose protocol
morning_dose <- create_protocol(
  name = "Morning dose",
  type = "oral",
  interval = "single",
  dose = 250,
  dose_unit = "mg"
)

# Add the administration to the schema
advanced_protocol <- add_administration(
  advanced_protocol,
  schema_name = "Daily dosing",
  administration = morning_dose
)

# Remove an administration from a schema
advanced_protocol <- remove_administration(
  advanced_protocol,
  schema_name = "Daily dosing",
  administration_name = "Morning dose"
)
}
