% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulation.R
\name{remove_formulation}
\alias{remove_formulation}
\title{Remove a formulation from a snapshot}
\usage{
remove_formulation(snapshot, formulation_name)
}
\arguments{
\item{snapshot}{a snapshot object.}

\item{formualtion_name}{a character vector of protocol name(s) to remove}
}
\value{
the updated snapshot without the specified formulations
}
\description{
Remove a formulation from a snapshot
}
\examples{
# Remove a single formulation
snapshot <- remove_formulation(snapshot, "Extended release tablet")

# Remove multiple formulations
snapshot <- remove_formulation(
  snapshot, 
  c("Immediate release solution", "Standard tablet")
)
}
