% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{remove_protocol}
\alias{remove_protocol}
\title{Remove a protocol from a snapshot}
\usage{
remove_protocol(snapshot, protocol_name)
}
\arguments{
\item{snapshot}{a snapshot object.}

\item{protocol_name}{a character vector of protocol name(s) to remove}
}
\value{
the updated snapshot without the specified protocols
}
\description{
Remove a protocol from a snapshot
}
\examples{
# Create a compound snapshot
midazolam <- compound("Midazolam")

# Create and add some protocols
protocol1 <- create_protocol("Protocol 1", "oral", "single", 100)
protocol2 <- create_protocol("Protocol 2", "iv", "24", 50)
midazolam <- add_protocol(midazolam, protocol1)
midazolam <- add_protocol(midazolam, protocol2)

# Remove a protocol from a snapshot
midazolam <- remove_protocol(midazolam, "Protocol 1")

# Remove multiple protocols at once
midazolam <- remove_protocol(midazolam, c("Protocol 1", "Protocol 2"))
}
