% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{set_compound_protocol}
\alias{set_compound_protocol}
\title{Set protocol for an already used compound from a \code{Simulation} object}
\usage{
set_compound_protocol(simulation, compound, protocol, formulation = list())
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}).}

\item{compound}{Name of the compound for which to set the protocol.}

\item{protocol}{either a protocol object (Protocol or AdvancedProtocol object) or a protocol name to be used for \code{compound}.
If a protocol name is used, the formulations should be set with each corresponding
formulation key required from the protocol.}

\item{formulation}{either a character vector of formulation names if a protocol object is given or a list of Formulation
key/name mapping for the chosen protocol if only its name is given.}
}
\value{
The updated \code{Simulation} object
}
\description{
This function set the protocol of the specified compounds.
}
\examples{
# Create a simulation first
sim <- create_simulation(
  simulation_name = "Protocol example",
  individual = "Adult male",
  victim = "Midazolam",
  perpetrators = c("Ketoconazole")
)

# Set protocol for a compound
sim <- set_compound_protocol(sim, "Midazolam", "Single oral dose")

# Set protocol with formulation
sim <- set_compound_protocol(
  sim,
  "Ketoconazole",
  "Multiple dose",
  formulation = list(list(Key = "Formulation 1", Name = "Tablet"))
)
}
