% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{set_compound_protocol}
\alias{set_compound_protocol}
\title{Set protocol for an already used compound from a \code{Simulation} object}
\usage{
set_compound_protocol(simulation, compound, protocol, formulation = list())
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}).}

\item{compound}{Name of the compound for which to set the protocol.}

\item{protocol}{Name of protocol used for \code{compound}.}

\item{formulation}{Formulation key/name mapping for the chosen protocol.}
}
\value{
The updated \code{Simulation} object
}
\description{
This function set the protocol of the specified compounds.
}
\examples{
# Set protocol for a compound
sim <- set_compound_protocol(sim, "Midazolam", "Single oral dose")

# Set protocol with formulation
sim <- set_compound_protocol(
  sim,
  "Ketoconazole",
  "Multiple dose",
  formulation = list(list(Key = "Formulation 1", Name = "Tablet"))
)
}
