% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{set_outputs}
\alias{set_outputs}
\title{Set outputs to a \code{Simulation} object}
\usage{
set_outputs(simulation, paths)
}
\arguments{
\item{simulation}{The \code{Simulation} object (as created by \code{create_simulation}).}

\item{paths}{path of the output to use for plotting}
}
\value{
The updated \code{Simulation} object.
}
\description{
Set the outputs to be used for plotting to the simulation.
This removes any previously set outputs.
}
