% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{Snapshot}
\alias{Snapshot}
\title{R6 Class Representing a Snapshot}
\description{
A snapshot is a JSON file that contains all the information about a model.
This class is used to import, access and manipulate the data from a snapshot.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source}}{The source of the snapshot (local file, url, ...)}

\item{\code{source_data}}{The input data of the snapshot}

\item{\code{version}}{version The version of the snapshot}

\item{\code{simulations_results}}{The simulation results of the snapshot}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{dynamic json representation of the snapshot object}

\item{\code{compounds}}{Access the compounds data from the snapshot.}

\item{\code{individuals}}{Access the individuals data from the snapshot.}

\item{\code{populations}}{Access the populations data from the snapshot.}

\item{\code{formulations}}{Access the formulations data from the snapshot.}

\item{\code{protocols}}{Access the protocols data from the snapshot.}

\item{\code{expression_profiles}}{Access the expression profiles data from the snapshot.}

\item{\code{observer_sets}}{Access the observer sets data from the snapshot.}

\item{\code{events}}{Access the events data from the snapshot.}

\item{\code{simulations}}{Access the simulations data from the snapshot.}

\item{\code{observed_data}}{Access the observed data from the snapshot.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Snapshot-new}{\code{Snapshot$new()}}
\item \href{#method-Snapshot-print}{\code{Snapshot$print()}}
\item \href{#method-Snapshot-export}{\code{Snapshot$export()}}
\item \href{#method-Snapshot-get_names}{\code{Snapshot$get_names()}}
\item \href{#method-Snapshot-add_protocol}{\code{Snapshot$add_protocol()}}
\item \href{#method-Snapshot-remove_protocol}{\code{Snapshot$remove_protocol()}}
\item \href{#method-Snapshot-run_simulations}{\code{Snapshot$run_simulations()}}
\item \href{#method-Snapshot-clone}{\code{Snapshot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-new"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-new}{}}}
\subsection{Method \code{new()}}{
Create a Snapshot object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$new(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{character string that is wether
\itemize{
\item a compound name from the list of available compounds \code{list_compounds()}.
\item a URL to a compound building block.
\item a Path to a local file.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Snapshot} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-print"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-print}{}}}
\subsection{Method \code{print()}}{
Pretty print the snapshot data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-export"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-export}{}}}
\subsection{Method \code{export()}}{
Export a DDI simulation to a JSON file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$export(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{character string that is the path to the output file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-get_names"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-get_names}{}}}
\subsection{Method \code{get_names()}}{
get_names the names of a field in the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$get_names(field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{the field to get the names from.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-add_protocol"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-add_protocol}{}}}
\subsection{Method \code{add_protocol()}}{
add a protocol to the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$add_protocol(protocol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocol}}{the protocol object to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-remove_protocol"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-remove_protocol}{}}}
\subsection{Method \code{remove_protocol()}}{
remove a protocol from the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$remove_protocol(protocol_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocol_name}}{name(s) of the protocol(s) to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-run_simulations"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-run_simulations}{}}}
\subsection{Method \code{run_simulations()}}{
run simulations defined in the snapshot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$run_simulations(path = NULL, exportPKML = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{character string that is the path to the output directory.}

\item{\code{exportPKML}}{logical that indicates if the PKML files should be exported.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
