% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{Snapshot}
\alias{Snapshot}
\title{R6 Class Representing a Snapshot}
\description{
A snapshot is a JSON file that contains all the information about a model.
This class is used to import, access and manipulate the data from a snapshot.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source}}{The source of the snapshot (local file, url, ...)}

\item{\code{data}}{The data of the snapshot}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{compounds}}{Access the compounds data from the snapshot.}

\item{\code{individuals}}{Access the individuals data from the snapshot.}

\item{\code{populations}}{Access the populations data from the snapshot.}

\item{\code{formulations}}{Access the formulations data from the snapshot.}

\item{\code{protocols}}{Access the protocols data from the snapshot.}

\item{\code{expression_profiles}}{Access the expression profiles data from the snapshot.}

\item{\code{observer_sets}}{Access the observer sets data from the snapshot.}

\item{\code{events}}{Access the events data from the snapshot.}

\item{\code{simulations}}{Access the simulations data from the snapshot.}

\item{\code{observed_data}}{Access the observed data from the snapshot.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Snapshot-new}{\code{Snapshot$new()}}
\item \href{#method-Snapshot-print}{\code{Snapshot$print()}}
\item \href{#method-Snapshot-compoundsNames}{\code{Snapshot$compoundsNames()}}
\item \href{#method-Snapshot-individualsNames}{\code{Snapshot$individualsNames()}}
\item \href{#method-Snapshot-populationsNames}{\code{Snapshot$populationsNames()}}
\item \href{#method-Snapshot-formulationsNames}{\code{Snapshot$formulationsNames()}}
\item \href{#method-Snapshot-protocolsNames}{\code{Snapshot$protocolsNames()}}
\item \href{#method-Snapshot-observerSetsNames}{\code{Snapshot$observerSetsNames()}}
\item \href{#method-Snapshot-eventsNames}{\code{Snapshot$eventsNames()}}
\item \href{#method-Snapshot-simulationsNames}{\code{Snapshot$simulationsNames()}}
\item \href{#method-Snapshot-observedDataNames}{\code{Snapshot$observedDataNames()}}
\item \href{#method-Snapshot-clone}{\code{Snapshot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-new"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-new}{}}}
\subsection{Method \code{new()}}{
Create a Snapshot object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$new(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{character string that is wether
\itemize{
\item a compound name from the list of available compounds \code{list_compounds()}.
\item a URL to a compound building block.
\item a Path to a local file.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Snapshot} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-print"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-print}{}}}
\subsection{Method \code{print()}}{
Pretty print the snapshot data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-compoundsNames"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-compoundsNames}{}}}
\subsection{Method \code{compoundsNames()}}{
Get the names of the compounds in the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$compoundsNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-individualsNames"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-individualsNames}{}}}
\subsection{Method \code{individualsNames()}}{
Get the names of the individuals in the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$individualsNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-populationsNames"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-populationsNames}{}}}
\subsection{Method \code{populationsNames()}}{
Get the names of the populations in the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$populationsNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-formulationsNames"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-formulationsNames}{}}}
\subsection{Method \code{formulationsNames()}}{
Get the names of the formulations in the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$formulationsNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-protocolsNames"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-protocolsNames}{}}}
\subsection{Method \code{protocolsNames()}}{
Get the names of the protocols in the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$protocolsNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-observerSetsNames"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-observerSetsNames}{}}}
\subsection{Method \code{observerSetsNames()}}{
Get the names of the observer sets in the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$observerSetsNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-eventsNames"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-eventsNames}{}}}
\subsection{Method \code{eventsNames()}}{
Get the names of the events in the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$eventsNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-simulationsNames"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-simulationsNames}{}}}
\subsection{Method \code{simulationsNames()}}{
Get the names of the simulations in the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$simulationsNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-observedDataNames"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-observedDataNames}{}}}
\subsection{Method \code{observedDataNames()}}{
Get the names of the observed data in the snapshot.
description
get_names the names of a field in the snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$observedDataNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Snapshot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Snapshot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Snapshot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
